// apps/renubu/src/lib/workflows/types.ts
export enum AccountPlan {
  INVEST = 'invest',
  EXPAND = 'expand',
  MANAGE = 'manage',
  MONITOR = 'monitor'
}

/**
 * Experience Level Enum
 */
export enum ExperienceLevel {
  JUNIOR = 'junior',
  MID = 'mid',
  SENIOR = 'senior',
  EXPERT = 'expert'
}